cupsd: If an administrative task is requested from a Snap, do extra checks
When CUPS is run as a Snap, we want to control by the interface the a
client Snap is using to connect to CUPS whether we allow
administrative tasks (create a new printer, cancel all jobs, ...) or
only user tasks (listing printers and capabilities, printing, cancel
the requester's own jobs, ...).

For this we identify at first whether an inquiry from a client coming
through a domain socket is administrative or not. We consider the task
administrative when it is only allowed due to the fact that the user
is in one of the groups listed in the "SystemGroup" directive in
cups-files.conf. This is the case when the task is authorized via
"require user @SYSTEM" line in the task's "limit" section of the
policy in cupsd.conf.

In this case we hook in another check, polling the PID of the client
process via peer creds and checking whether this PID is under AppArmor
confinement and whether the confinement comes from a Snap, and from
which Snap.

Planned, but not yet implemented in this commit, is to check whether
the client Snap plugs the "cups-control" interface and then allow the
task, otherwise to deny it. In snapd we will replace the curremt
"cups-control" interface by two interfaces then, "cups" which only
allows user tasks and "cups-control" which will also allow
administrative tasks using this patch.

Client Snaps plugging "cups" will connect automatically, clients
plugging "cups-control" will require manual connection or permission
for automatic connection in the Snap Store.

See

https://forum.snapcraft.io/t/interface-request-cups-control-on-cups-snap-and-including-d-bus/

The current commit simply denies administrative operations requested
by Snaps.