cupsd: Added check whether client Snap connects via "cups-control" interface
Now we have a complete working model for our patch which applies the
same criteria to allowing administrative CUPS tasks as PulseAudio
applies to allow audio recording. See the patch for PulseAudio as Snap
here:

https://git.launchpad.net/ubuntu/+source/pulseaudio/tree/debian/patches/0700-modules-add-snappy-policy-module.patch

After determining the name of the client Snap via AppArmor we use the
snapd-glib library functions to first poll the general data record of
the Snap and with this check whether it is under classic
confinement. If so, we allow the operation.

If our client Snap is NOT under classic confinement, we poll its
connections but only its plugging of "cups-control". If we find it
plugging "cups-control" we allow the operation, otherwise we deny it.

The work still needed is improving the error handling, as any error on
the way makes the operation be allowed, destroying the connection to
snapd when done, to avoid memory leaks, and some code polishing.